Major update and expansion of GainMatcher. Moved from linear to quadratic fit and therefore providing three coefficients instead of two. The code uses four peaks instead of two. Manual mode changed so that shift-click identies the centroid of each peak instead of selecting the limits of a search region. Auto mode does not require any shift-click but automates the search regions. Added a button to download the coeffients as a text file in the GRSISort Cal file format where the title includes the run number to identify the data used for the fits.
Major expansion of the intelligence built into the Guess peaks function which is made possible by there being four peaks rather than two. This allows consistency checks for the gain between the different peaks identified and selection of the peaks which best fit the energies provided. The fitting is now quite reliable and rarely fails to guess the correct peaks. The subsequent peak fitting does fail sometimes becasue of particularly narrow or weak peaks. If only three peaks are fitted correctly then the regression of the quadratic fit is performed on only the three fitted peaks instead of four. This is highlighted in the table because one peak channel is identified as NaN.